refactor: changing the 'value' data type from BigDecimal to Double.

+++ src/main/java/br/com/phlimadev/itau_challenge/transaction/TransactionDTO.java
@@ -4,13 +4,12 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Past;
 import jakarta.validation.constraints.PositiveOrZero;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record TransactionDTO(
         @NotNull(message = "O campo 'valor' deve ser preenchido.")
         @PositiveOrZero(message = "A transação DEVE ter valor igual ou maior que 0 (zero)")
-        BigDecimal valor,
+        Double valor,
         @NotNull(message = "O campo 'dataHora deve ser preenchido'")
         @Past(message = "A transação DEVE ter acontecido a qualquer momento no passado")
         OffsetDateTime dataHora
